// api/models/bookmark.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Audiobookshelf API
 * Audiobookshelf API with autogenerated OpenAPI doc
 * OpenAPI spec version: 0.1.0
 */
import type { CreatedAt } from './createdAt';
import type { LibraryItemId } from './libraryItemId';
import type { Time } from './time';

export interface Bookmark {
  libraryItemId?: LibraryItemId;
  /** The title of the bookmark. */
  title?: string;
  time?: Time;
  createdAt?: CreatedAt;
}