// api/models/user.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Audiobookshelf API
 * Audiobookshelf API with autogenerated OpenAPI doc
 * OpenAPI spec version: 0.1.0
 */
import type { Bookmark } from './bookmark';
import type { CreatedAt } from './createdAt';
import type { IsActive } from './isActive';
import type { IsLocked } from './isLocked';
import type { LastSeen } from './lastSeen';
import type { LibraryId } from './libraryId';
import type { LibraryItemId } from './libraryItemId';
import type { MediaProgress } from './mediaProgress';
import type { Permissions } from './permissions';
import type { Tags } from './tags';
import type { Token } from './token';
import type { UserId } from './userId';
import type { UserType } from './userType';
import type { Username } from './username';

/**
 * A user object which includes the user's ID, username, type, token, and media progress.
 */
export interface User {
  id?: UserId;
  username?: Username;
  type?: UserType;
  token?: Token;
  /** The user's progress of media. */
  mediaProgress?: MediaProgress[];
  /** The IDs of series to hide from the user's "Continue Series" shelf. */
  seriesHideFromContinueListening?: LibraryItemId[];
  /** The user's bookmarks. */
  bookmarks?: Bookmark[];
  lastSeen?: LastSeen;
  isActive?: IsActive;
  isLocked?: IsLocked;
  createdAt?: CreatedAt;
  permissions?: Permissions;
  /** The IDs of libraries the user has access to. */
  librariesAccessible?: LibraryId[];
  itemTagsSelected?: Tags;
}